refactor: lazy-load FullCalendar via next/dynamic (P0-3)
